sauti_cms/requirements.txt
Django==5.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
python-decouple==3.8
drf-spectacular==0.28.0
Pillow==11.3.0
gunicorn==23.0.0
whitenoise==6.9.0
cryptography==44.0.0
